Update ci.yml to use ubuntu 22.04 to fix subsequent t64 package installation errors
Ubuntu 24.04 LTS solves the Year 2038 problem 2.0k that existed on armhf. More than a thousand packages have been updated to handle time using a 64-bit value rather than a 32-bit one, making it possible to handle times up to 292 billion years in the future.

That is the reason for the new name libasound2t64 with t64 at the end.